Moved files in include/ to src/ and removed the include dir.
