+++ src/mathfunctions.h
